+++ examples/resources/morpheus_environment/resource.tf
@@ -1,6 +1,6 @@
-resource "morpheus_environment" "sample" {
+resource "morpheus_environment" "tf_example_environment" {
   active      = true
-  code        = "staging"
-  description = "Staging"
-  name        = "Staging"
+  code        = "tfexample"
+  description = "Terraform Example"
+  name        = "tfexample"
 }